Add support for specifying basic authentication
Motivation:

ownCloud and NextCloud support WebDAV access to shares; however, both
require that the "share ID" is passed as the username.  If the share is
public then the HTTP basic authentication password should be the empty
string, otherwise the share password should be used.

Modification:

Add support for specifying the username (and optionally a password)
through the URL; e.g,.

    https://NJtdJwYLdfgyGN8:@nubes.helmholtz-berlin.de/public.php/webdav/

The generated metalink file will then contain the userinfo part in the
url elements, as supplied in the argument.

Result:

It's now possible to generate a metalink file for the WebDAV endpoint of
an ownCloud or NextCloud share.