Improve data URL discovery.
This patch makes two changes.

First, it avoids hard-coding the IDS URL.  Instead, the IDS URL is
discovered by issuing an HTTP HEAD request against the ICAT+ API,
following any redirections.

Second, it makes this redirection optional.  By default, the ICAT+ API
link is used as the data URL.  If the '-f' option is specified then a
HEAD request is used to discover the true URL from which data may be
transferred.