Add suppress file size option
Motivation:

Aria2c will send a Range header if the file size is known.  This causes
a problem for ESRF, as the server doesn't know the file's size,
resulting in the server responding with a 400 Bad Request response.

Modification:

Add option to suppress recording the file size.

Result:

Data may be transferred.