Add DOI as 'metalink:origin'
Motivation:

RFC 5854 says that a Metalink file SHOULD contain a `metalink:origin`
element.

Modification:

Use the DOI as the origin of the metalink file.  This is somewhat
contentious, as the semantics of metalink:origin is either to give the
metalink file a unique identity, or to provide a way for a client to
obtain an updated version of the metalink.  The latter would be possible
if the D3A proposal is adopted.

Normally, a DOI resolves (via the DOI proxy) to a landing page;
therefore, specifying a URL (e.g., `https://doi.org/10.xxx`) might
result in a program receiving a confusing response.  To reduce this
risk, the URN form (`doi:10.xxx`) is used.

Result:

The ESRF Metalink files now include a metalink:origin element