Refactor file walking logic
Motivation:

The current code mixes the logic for walking/visiting all files at a
WebDAV endpoint via recursion with the code that emits a metalink
description of those files.

In the interests of "seperation of concern" and supporting possible
future extensions, these two activities should be seperated.

Modification:

Create an abstract walker class that is responsible for visiting all
files and directories underneath a WebDAV endpoint.

The existing metalink emitting code is now available as a separate
class.

Result:

It is now easier to extend the class to handle other use-cases.